requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
python-docx==1.1.2
python-dateutil==2.9.0
openpyxl==3.1.2
google-auth==2.29.0
google-api-python-client==2.129.0
psycopg2-binary==2.9.9    # Para PostgreSQL/Supabase en producción
alembic==1.13.1            # Migraciones de BD
